Completed functionality of zombie game.
The heart of the problem is solved in the kill-all method of Zombies
which produces a pair of undead and dead sets of zombies.  It does so
by shuffling any zombies in the undead set that are touching any other
zombie in either the dead or undead sets.  The world% eat-brains
method uses kill-all to compute the world resulting from zombies
eating other zombies.

Introduced the location interface, which includes anything that has a
posn method.  Removed mouse% in favor of using posn% to represent
mice.  Wrote some tests and cleaned things up.