Refactored world to be representation independent of zombies.
Introduced the Hoard interface which represents the collection of
undead and dead zombies.  The world now has a single hoard and sends
it messages.  Currently the hoard is implemented as a pair of lists of
zombies, but if we changed it to a single list of zombies that are
either dead or undead, the world program would not need to change.

The world% implementation is now quite nice.  The eat-brains method
went away, as did advance.  The stop-when method has a nicer
definition.  So does to-draw.